tests/wpt/web-platform-tests/workers/constructors/SharedWorker/global-members.html

<!doctype html>
<title>members of SharedWorkerGlobalScope</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async_test(t => {
  const worker = new SharedWorker('global-members.js');
  worker.port.onmessage = t.step_func_done(e => {
    assert_equals(e.data, '');
  });
}, 'Test if global members exist in a shared worker');
</script>
